Rewrite normalize-url
