+++ .eslintignore
@@ -2,4 +2,3 @@ dist
 generated-client
 lib
 node_modules
-src/utils/normalize-url
