+++ .github/dependabot.yml
@@ -6,9 +6,6 @@ updates:
     interval: weekly
   open-pull-requests-limit: 10
   rebase-strategy: disabled
-  ignore:
-    - dependency-name: normalize-url
-      update-types: [ version-update:semver-major ]
 
 - package-ecosystem: github-actions
   directory: /
