--- src/utils/normalize-url/package.json
@@ -1,52 +1,0 @@
-{
-	"name": "normalize-url",
-	"version": "7.0.3",
-	"description": "Normalize a URL",
-	"license": "MIT",
-	"repository": "sindresorhus/normalize-url",
-	"funding": "https://github.com/sponsors/sindresorhus",
-	"author": {
-		"name": "Sindre Sorhus",
-		"email": "sindresorhus@gmail.com",
-		"url": "https://sindresorhus.com"
-	},
-	"type": "module",
-	"exports": "./index.js",
-	"engines": {
-		"node": ">=12.20"
-	},
-	"scripts": {
-		"test": "ava && tsd"
-	},
-	"files": [
-		"index.js",
-		"index.d.ts"
-	],
-	"keywords": [
-		"normalize",
-		"url",
-		"uri",
-		"address",
-		"string",
-		"normalization",
-		"normalisation",
-		"query",
-		"querystring",
-		"simplify",
-		"strip",
-		"trim",
-		"canonical"
-	],
-	"devDependencies": {
-		"ava": "^4.0.1",
-		"c8": "^7.11.0",
-		"tsd": "^0.19.1",
-		"xo": "^0.47.0"
-	},
-	"c8": {
-		"reporter": [
-			"text",
-			"lcov"
-		]
-	}
-}
