+++ tsconfig.json
@@ -1,7 +1,6 @@
 {
   "extends": "@tsconfig/recommended/tsconfig.json",
   "compilerOptions": {
-    "allowJs": true,
     "target": "ES5",
     "declaration": true,
     "noImplicitAny": true,
